fix flyway migration

+++ src/main/resources/db/migration/V2__Create_workout_plans_table.sql
@@ -1,6 +1,6 @@
 CREATE TABLE IF NOT EXISTS workout_plans (
                                              id BIGINT AUTO_INCREMENT PRIMARY KEY,
-                                             user_id INT NOT NULL,
+                                             user_id BIGINT NOT NULL,
                                              title VARCHAR(255) NOT NULL,
                                              description TEXT,
                                              created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP NOT NULL,
